🏗️ Architect: Refactor iomap_otbm and organize UI files
- Extracted Item OTBM serialization logic from `source/io/iomap_otbm.cpp` to `source/game/item_otbm.cpp`.
- Organized `source/ui/` by moving window files to `source/ui/windows/`.
- Organized `source/ui/` by moving menubar files to `source/ui/menubar/` and toolbar files to `source/ui/toolbar/`.
- Updated `CMakeLists.txt` and `#include` directives to reflect file moves.
- Fixed missing includes in `source/io/filehandle.h` and `source/game/items.h` to ensure standalone compilation.
- Verified build with `ninja` and `cmake`.
