+++ source/ui/menubar/map_actions_handler.cpp
@@ -1,7 +1,7 @@
 #include "ui/menubar/map_actions_handler.h"
 #include "ui/gui.h"
 #include "ui/dialog_util.h"
-#include "ui/find_item_window.h"
+#include "ui/windows/find_item_window.h"
 #include "editor/editor.h"
 #include "editor/operations/clean_operations.h"
 #include "editor/operations/search_operations.h"
