+++ source/ui/menubar/search_handler.cpp
@@ -3,9 +3,9 @@
 
 #include "ui/gui.h"
 #include "ui/dialog_util.h"
-#include "ui/find_item_window.h"
-#include "ui/result_window.h"
-#include "ui/map_window.h"
+#include "ui/windows/find_item_window.h"
+#include "ui/windows/result_window.h"
+#include "ui/windows/map_window.h"
 #include "app/preferences.h"
 #include "editor/editor.h"
 #include "editor/operations/search_operations.h"
